requirements.txt
Flask==2.3.3
Pillow==9.4.0
numpy==1.24.4
opencv-python==4.8.0.76
keras==2.12.0
tensorflow==2.12.0
